Removed old SVG X/Y chart and replace with the Canvas one (#88)
* Removed old SVG X/Y chart and replace with the Canvas one

* Remove old pointSeries
