+++ packages/perspective-viewer-d3fc/src/less/chart.less
@@ -30,7 +30,7 @@
         & d3fc-canvas.plot-area {
             display: none;
         }
-        &.d3_xy_scatter_canvas {
+        &.d3_xy_scatter {
             & d3fc-canvas.plot-area {
                 display: block;
             }
