+++ packages/perspective-viewer-d3fc/test/results/results.json
@@ -1,41 +1,41 @@
 {
-    "scatter.html/displays visible columns.": "92c93c38904ab90cd9fc9a648d68d3de",
+    "scatter.html/displays visible columns.": "8721d9a2911ed5e22441eb8f81918d9b",
     "line.html/shows a grid without any settings applied.": "d62b69c1b351faeecd0992c0d200702f",
-    "line.html/pivots by a row.": "ea25be25e522ef15f21bc799a74f1700",
-    "line.html/pivots by two rows.": "1a618ac9e3682a065b4e775515700dc1",
-    "line.html/pivots by a column.": "fd85b140660c1cbc410a84bf0977d611",
-    "line.html/pivots by a row and a column.": "8916360f62868ec430390b7c6a07d37d",
-    "line.html/pivots by two rows and two columns.": "d121555d3a9093544277d4c781f72816",
-    "line.html/sorts by a hidden column.": "46e24dbc61cb485079e6750ffecda381",
-    "line.html/sorts by a numeric column.": "37ab6e0ac05f31752fa869ee2dcf52f3",
-    "line.html/filters by a numeric column.": "92e6ec76042209973c2de43db1a4723d",
-    "line.html/filters by a datetime column.": "d4986c73717e9fa792160fe4ba906e95",
-    "line.html/highlights invalid filter.": "18fa430202ff53e0f3c3dabb7df775e4",
-    "line.html/sorts by an alpha column.": "c8708e4e129dcdc1f19374e0c8b253b9",
-    "line.html/displays visible columns.": "a7fbdd3af7e39efce32eeca573c2dba2",
+    "line.html/pivots by a row.": "354e25ef3265d4f8003ca35647199ca3",
+    "line.html/pivots by two rows.": "fd40a4bd1236374c116c4387f7f2a5ae",
+    "line.html/pivots by a column.": "7cccaf4c6292f63b5a72f9e4c215d54b",
+    "line.html/pivots by a row and a column.": "857b187307c377c3c2cde36481a4e411",
+    "line.html/pivots by two rows and two columns.": "60111f869482926cd56061cdbfedfe4b",
+    "line.html/sorts by a hidden column.": "b4814e6751af51f79f026b3119dd08c5",
+    "line.html/sorts by a numeric column.": "9a7996cdfd69dc89fc80232a302c388a",
+    "line.html/filters by a numeric column.": "e681ad1e4b153991c2ab6ffe1e96441f",
+    "line.html/filters by a datetime column.": "4c11d78604c2226745b3b97637f2da3f",
+    "line.html/highlights invalid filter.": "900b78d70684b179da5452e0956ff88f",
+    "line.html/sorts by an alpha column.": "d99e768606bb41e1bbf597f36baf6eed",
+    "line.html/displays visible columns.": "12dbe61a2f207f4cbd875d85b5aff669",
     "bar.html/shows a grid without any settings applied.": "d00de230f6aed149a6619a75eae91924",
-    "bar.html/pivots by a row.": "881cac48dae1c2dda037365c67194a5f",
-    "bar.html/pivots by two rows.": "c942707311ca6275207529c159bca7e9",
-    "bar.html/pivots by a column.": "4d6bb531645cf714de982f7bd7810b76",
-    "bar.html/pivots by a row and a column.": "611e7684024edf99621e6d568c1fdc73",
-    "bar.html/pivots by two rows and two columns.": "9cbecc887b723916cda2f87e3e2be419",
-    "bar.html/sorts by a hidden column.": "35e2dbbc7966fa6ba2dc8e31e7429e7f",
-    "bar.html/sorts by a numeric column.": "3887b4cd90cc9c002f0a0f9722d74883",
-    "bar.html/filters by a numeric column.": "3ed5b1817cb0b114f9403248f3ceb2fc",
-    "bar.html/filters by a datetime column.": "8008cdbd7ab52d0c0904f62fd9782ea7",
-    "bar.html/highlights invalid filter.": "42dae65f09491f2630962bc18d99ddfa",
-    "bar.html/sorts by an alpha column.": "c08b9db323217d9ddbfcdb5bc411172c",
-    "bar.html/displays visible columns.": "3b03df696716f2a1349840d581c5aed1",
-    "scatter.html/shows a grid without any settings applied.": "c66e7c845fd48ee2b4e07088686e2d6a",
-    "scatter.html/pivots by a row.": "a1884559b36693e9b12d8923c4ec2ab0",
-    "scatter.html/pivots by two rows.": "899da6085c750e28278818b920bda326",
-    "scatter.html/pivots by a column.": "de55de9a2b7d0475575a35c927c1e5b4",
-    "scatter.html/pivots by a row and a column.": "50e861f0b90d179322ea636c5cbf61f3",
-    "scatter.html/pivots by two rows and two columns.": "b4e53b4532a1e91ca0d6952cf1f17150",
-    "scatter.html/sorts by a hidden column.": "2c35f52f3678c9f35f297e0077d773e7",
-    "scatter.html/sorts by a numeric column.": "ca0936a0844c028cfb20696ecf51eb40",
-    "scatter.html/filters by a numeric column.": "18f5c8e7292448d153ff9fd18f66aad3",
-    "scatter.html/filters by a datetime column.": "3bdd41ad10fa4a9d79d9c85773a2e300",
-    "scatter.html/highlights invalid filter.": "c9765fbcbbe6b8cee8257d25b26ec7db",
-    "scatter.html/sorts by an alpha column.": "59f05e4a08b06ead82019df65ee3f56c"
+    "bar.html/pivots by a row.": "957891b1b2df1d97af813ba1a68ce0da",
+    "bar.html/pivots by two rows.": "c06b518f3af00f0692fdda004dcc3e97",
+    "bar.html/pivots by a column.": "8bf3bbe3c5afd3ab5f6cc8fcc5e25a45",
+    "bar.html/pivots by a row and a column.": "eda5147299cef72e9995e7ad9e74b50d",
+    "bar.html/pivots by two rows and two columns.": "54942e792e5ecf51d33699aaec6b5434",
+    "bar.html/sorts by a hidden column.": "38b74a4304e90d55bc6a679fe68168ce",
+    "bar.html/sorts by a numeric column.": "588d8fecb67c4c521094ca16a26392a5",
+    "bar.html/filters by a numeric column.": "faec09764578964c714801a4723226aa",
+    "bar.html/filters by a datetime column.": "1ab6112ffb47eccfe6fb4083a2f99602",
+    "bar.html/highlights invalid filter.": "aa3417ce8fd895110abcfb6a17a7e41a",
+    "bar.html/sorts by an alpha column.": "cf1bf3c5a9dd6d5771067b8b500872c9",
+    "bar.html/displays visible columns.": "b6fb208f5e561afeb470d041670cdaa0",
+    "scatter.html/shows a grid without any settings applied.": "0424867aec4a72628c3a09af275a9c61",
+    "scatter.html/pivots by a row.": "970eb4ffc6a856826f1c4b64c8df0703",
+    "scatter.html/pivots by two rows.": "936e368b628a37c08bfca039642b7acd",
+    "scatter.html/pivots by a column.": "5e0884b058ffb75e3ff6d6d7d8bd9d82",
+    "scatter.html/pivots by a row and a column.": "5066f0ef9f528b4ae40b84f1b85cf05f",
+    "scatter.html/pivots by two rows and two columns.": "14370acb9556204599b71177aae0bd7d",
+    "scatter.html/sorts by a hidden column.": "bb95df5c3912f0c2f93357bfaaa77ced",
+    "scatter.html/sorts by a numeric column.": "77288d5574d94af6e63fda631efb9b77",
+    "scatter.html/filters by a numeric column.": "810731d627d0cca8c4e616c0eb9fda53",
+    "scatter.html/filters by a datetime column.": "8ecc3bdf51762974b9a9784d7468f1ae",
+    "scatter.html/highlights invalid filter.": "a049b73d94bc919678e121be213ece07",
+    "scatter.html/sorts by an alpha column.": "42fce2942e4c83e6a5ded79097c90571"
 }
